Make css-comb work for regions
csscomb doesn't follow unix' conventions of taking stdin/pooping stdout, so it's not as straightforward to program it in elsip. So, we create a temporary file for the buffer to comb, copy the buffer over (because buffer might not always be visiting a file), comb it, and read the content back.

For the region, since csscomb "requires" a file, we create a new buffer with the region, and comb it with above method. But csscomb generates some uglyass css, so I am making beauty versions of it next, using web-beautify-css