Solve web-beautify-format-buffer problem
We were using ~web-beautify-format-buffer~ to format the combed css,
which was failing with an error:

progn: Wrong number of arguments: #[(program) "Áed#" [program web-beautify-format-region] 4 ("/Users/channikhabra/.emacs.d/elpa/web-beautify-20160410.1005/web-beautify.elc" . 2362)], 2

It should've been working as of the docs of web-beautify-format-buffer
function, but it isn't. However, ~web-beautify-css-buffer~ seem to be
doing the job just fine, and we're now using that.